LayoutTests/svg/dom/SVGTests-expected.txt

This test checks the SVGTests API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check the requiredFeatures, requiredExtensions and systemLanguage attributes
PASS foreignObject.requiredFeatures instanceof SVGStringList is true
PASS foreignObject.requiredExtensions instanceof SVGStringList is true
PASS foreignObject.systemLanguage instanceof SVGStringList is true

Check the hasExtension function
FAIL foreignObject.hasExtension('http://www.w3.org/1998/Math/MathML') should be true. Threw exception TypeError: foreignObject.hasExtension is not a function. (In 'foreignObject.hasExtension('http://www.w3.org/1998/Math/MathML')', 'foreignObject.hasExtension' is undefined)
FAIL foreignObject.hasExtension('http://www.w3.org/1999/xhtml') should be true. Threw exception TypeError: foreignObject.hasExtension is not a function. (In 'foreignObject.hasExtension('http://www.w3.org/1999/xhtml')', 'foreignObject.hasExtension' is undefined)
FAIL foreignObject.hasExtension('') should be false. Threw exception TypeError: foreignObject.hasExtension is not a function. (In 'foreignObject.hasExtension('')', 'foreignObject.hasExtension' is undefined)
FAIL foreignObject.hasExtension('unknownExtension') should be false. Threw exception TypeError: foreignObject.hasExtension is not a function. (In 'foreignObject.hasExtension('unknownExtension')', 'foreignObject.hasExtension' is undefined)
FAIL foreignObject.hasExtension('HTTP://WWW.W3.ORG/1999/XHTML') should be false. Threw exception TypeError: foreignObject.hasExtension is not a function. (In 'foreignObject.hasExtension('HTTP://WWW.W3.ORG/1999/XHTML')', 'foreignObject.hasExtension' is undefined)
FAIL foreignObject.hasExtension('http://www.w3.org/1998/Math/MathML http://www.w3.org/1999/xhtml') should be false. Threw exception TypeError: foreignObject.hasExtension is not a function. (In 'foreignObject.hasExtension('http://www.w3.org/1998/Math/MathML http://www.w3.org/1999/xhtml')', 'foreignObject.hasExtension' is undefined)
PASS successfullyParsed is true

TEST COMPLETE
